Consistancy for getSurfacpePoint : Pipe problem fix   This is PART of the problem with pipelines     This patch fixes the inconsistancy with getSurfacePoint     It is part of three patches needed to fix the pipeline problem     The other two will by getLayerSurf and getLayerString.
It is definate that these are causeing errors it is not known what else
remains

+++ Model/essBuildInc/ButterflyModerator.h
@@ -74,7 +74,7 @@ class ButterflyModerator :
   virtual ButterflyModerator* clone() const;
   virtual ~ButterflyModerator();
   
-  virtual Geometry::Vec3D getSurfacePoint(const size_t,const size_t) const;
+  virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
   virtual int getLayerSurf(const size_t,const size_t) const;
   virtual std::string getLayerString(const size_t,const size_t) const;
   virtual int getCommonSurf(const size_t) const;
