+++ Model/essBuildInc/ConicModerator.h
@@ -82,7 +82,7 @@ class ConicModerator : public constructSystem::ModBase
   /// Access to hydrogen region
   virtual int getMainBody() const { return HCell; }
 
-  virtual Geometry::Vec3D getSurfacePoint(const size_t,const size_t) const;
+  virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
   virtual int getLayerSurf(const size_t,const size_t) const;
   virtual std::string getLayerString(const size_t,const size_t) const;
