+++ Model/essBuildInc/MidWaterDivider.h
@@ -79,7 +79,7 @@ class MidWaterDivider :
   virtual MidWaterDivider* clone() const;
   virtual ~MidWaterDivider();
 
-  virtual Geometry::Vec3D getSurfacePoint(const size_t,const size_t) const;
+  virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
   virtual std::string getLayerString(const size_t,const size_t) const;
   virtual int getLayerSurf(const size_t,const size_t) const;
   void createAll(Simulation&,const attachSystem::FixedComp&,
