+++ Model/photonInc/CylContainer.h
@@ -77,7 +77,7 @@ class CylContainer :
   /// Accessor to the main H2 body
   virtual int getMainBody() const { return cylIndex+1; }
 
-  virtual Geometry::Vec3D getSurfacePoint(const size_t,const size_t) const;
+  virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
   virtual int getLayerSurf(const size_t,const size_t) const;
   virtual int getCommonSurf(const size_t) const;
   virtual std::string getLayerString(const size_t,const size_t) const;
