+++ Model/t1UpgradeInc/CH4Layer.h
@@ -3,7 +3,7 @@
  
  * File:   t1UpgradeInc/CH4Layer.h
  *
- * Copyright (c) 2004-2015 by Stuart Ansell
+ * Copyright (c) 2004-2016 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -76,7 +76,7 @@ class CH4Layer : public constructSystem::ModBase
   virtual ~CH4Layer();
 
   virtual Geometry::Vec3D 
-    getSurfacePoint(const size_t,const size_t) const;
+    getSurfacePoint(const size_t,const long int) const;
   virtual int 
     getLayerSurf(const size_t,const size_t) const;
   virtual std::string
