+++ Model/t1UpgradeInc/H2Section.h
@@ -3,7 +3,7 @@
  
  * File:   t1UpgradeInc/H2Section.h
  *
- * Copyright (c) 2004-2015 by Stuart Ansell
+ * Copyright (c) 2004-2016 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -79,7 +79,7 @@ class H2Section : public constructSystem::ModBase
 
   virtual void addToInsertChain(attachSystem::ContainedComp&) const;
 
-  virtual Geometry::Vec3D getSurfacePoint(const size_t,const size_t) const;
+  virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const;
   virtual int getLayerSurf(const size_t,const size_t) const;
   virtual std::string getLayerString(const size_t,const size_t) const;
