+++ Model/t1UpgradeInc/SplitInner.h
@@ -3,7 +3,7 @@
  
  * File:   t1UpgradeInc/SplitInner.h
  *
- * Copyright (c) 2004-2015 by Stuart Ansell
+ * Copyright (c) 2004-2016 by Stuart Ansell
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -63,7 +63,7 @@ class SplitInner : public ts1System::CH4Layer
   virtual ~SplitInner();
 
   virtual Geometry::Vec3D 
-    getSurfacePoint(const size_t,const size_t) const;
+    getSurfacePoint(const size_t,const long int) const;
   virtual int 
     getLayerSurf(const size_t,const size_t) const;
   virtual std::string
