+++ System/attachCompInc/LayerComp.h
@@ -50,7 +50,7 @@ class LayerComp
 			       const Geometry::Vec3D&,
 			       const Geometry::Vec3D&) const;
 
-  virtual Geometry::Vec3D getSurfacePoint(const size_t,const size_t) const =0;
+  virtual Geometry::Vec3D getSurfacePoint(const size_t,const long int) const =0;
   virtual int getLayerSurf(const size_t,const size_t) const =0;
   virtual std::string getLayerString(const size_t,const size_t) const =0;
   /// Access to common divider surface
